Add route prefixing so the auth server can operate on a prefix
Needed this as we are moving the auth server to be hosted under the
same domain as the register api, but on a prefixed route.

This change allows URL_PREFIX to be set which will cause ALL urls
to be prefixed with this string. Eg: URL_PREFIX is set to "/prefix"
which will mean the token url can be reached here: "/prefix/oauth/token"