Only use Rack::Attack if running in PaaS
Currently the auth server is not starting up properly in the new AWS environment because the rack_attack_config is looking for a Redis that isn't there.

This adds a helper to detect if the app is running in GOV.UK PaaS or elsewhere, and only tries to connect to a real Redis if in PaaS.

We don't currently have Redis in our new AWS environment, and will manage rate limiting, etc. at WAF/CDN level rather than by the application.